refactor: Refactor code to extract JSON object from response
The code has been refactored to include a regular expression that extracts the JSON object from the response. This ensures that the extracted JSON object is returned as the response. If the JSON object cannot be extracted, an error is thrown. This change improves the reliability and accuracy of the response handling.